refactor(kitchenpos/menus): move common domain models to domain package
