+++ src/main/java/kitchenpos/common/tobe/domain/DisplayedName.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.tobe;
+package kitchenpos.common.tobe.domain;
 
 import java.util.Objects;
