+++ src/main/java/kitchenpos/common/tobe/domain/Quantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.tobe;
+package kitchenpos.common.tobe.domain;
 
 public class Quantity {
