+++ src/main/java/kitchenpos/menus/tobe/domain/MenuGroup.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.tobe.domain;
 
-import kitchenpos.common.tobe.DisplayedName;
+import kitchenpos.common.tobe.domain.DisplayedName;
 
 import java.util.UUID;
