+++ src/test/java/kitchenpos/common/tobe/DisplayedNameTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.common.tobe;
 
+import kitchenpos.common.tobe.domain.DisplayedName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
